// src/server/shared/lib/pncp/models/RecuperarDadosDocumentoContratoDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RecuperarDadosDocumentoContratoDTO = {
  uri?: string;
  url?: string;
  cnpj?: string;
  anoCompra?: number;
  sequencialCompra?: number;
  dataPublicacaoPncp?: string;
  titulo?: string;
  sequencialDocumento?: number;
  tipoDocumentoNome?: string;
};